Decompose main.py into utils
